[preflight] Add DevRel team as OWNERS
As the team looks to resolve some bugs and feature requests in
preflight. It makes sense to have DevRel claim ownership of this plugin
and experience especially as it is critical for on-boarding of new
fuchsia developers (as documented in fuchsia.dev).

Change-Id: Iabb2b716fd74151f8cc771191f02065392f0d282
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637411
Fuchsia-Auto-Submit: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Matt Boetger <boetger@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>