[blobfs] remove fuchsia.blobfs c fidl bindings
Only a couple of places imported them and those places didn't even use
them. Remove the imports and the annotation that generates the bindings.

This also adds discoverable to the fuchsia.blobfs.Blobfs protocol. After
blobfs is a component, this protocol is going to be placed in the
outgoing directory instead of being multiplexed over directory
connections, so I added it in since we will eventually need it and we
are modifying it anyway.

Bug: 90556
Test: fx build
Change-Id: I59da9d277fba362981d3c3b3915d7f23e0610170
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633774
Reviewed-by: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>