[starnix] Parse BufferCollectionToken from ImageFile
ImageFile also contains a BufferCollectionImportToken that needs to be
sent from the client to the wayland bridge. This change adds the code
that parses the token out of ImageFile, and adds an error log when the
file type is invalid.

Change-Id: I10737d8dd50fcaa217ad49b0717b01e7af337ba9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637581
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>