[fdr] Add more verbose logging.
With workstation we now have two zxcrypt partitions that need shredding
and its a bit tricky both to determine which partition failed when an
error occurs and to verify that both were successfully found and
shredded in the success case.

This just adds to more log lines to make things clearer.

Change-Id: I1a81138e6179fd85b3cd8cbff78872736e1940a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637742
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>