[display][rust] Mechanism to apply a layer configuration
Introduced a mechanism to declaratively describe a display layer
configuration and apply it to the display driver with validation. Not
all cursor and primary layer parameters are currently supported (notably
transforms, blending, and fences) but sufficient support is there to
build a primary layer with a shared image.

Change-Id: I907ebf90409bb97735307357576d9427226cff8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611802
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>