[camera][gym] Do not stop on PEER_CLOSED (buf col)
This CL is a re-land of fxr/629444:

Continue on BufferCollection PEER_CLOSED "error" because sysmem could
simply be reflecting a PEER_CLOSED by another client, and if so, then
this is a case where there are no real errors but merely signaling a
tear down.

The new assumption is to tear down the BufferCollection as camera-gym
knows it. In this code path, there is also an allowance for potential
race condition where camera-gym might a tear down might have occurred
already.

All errors for BufferCollection are now treated this way.

Bug: b/209523797
Test: camera-gym run consistently >10 hours without camera-gym exit.
      Verify that, upon disconnect, the BufferCollection unbound case
      executes.
Change-Id: I5a6532f9c8a263ff141d3bcca124a4e2643efb24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637416
Reviewed-by: Matt Sandy <msandy@google.com>
Fuchsia-Auto-Submit: Ernest Hua <ernesthua@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>