[thermal] Associate ThermalLoad with sensor path
The Power Manager uses ThermalLoad to represent the severity of thermals
in the system at any given time. Today, ThermalLoad doesn't incorporate
any knowledge about *which* sensor is causing the thermal load. Starting
with Nelson, we need to monitor multiple sensors separately, and
understand which sensor is contributing to the thermal load.

This change adds support within Power Manager to track the sensor
associated with a ThermalLoad value. With this change, the
ThermalLimiter now has access to see which sensor is causing the thermal
load change. However, it does not yet leverage that extra information.
Future CLs will build upon this added support.

Bug: fxbug.dev/90157
Test: fx test power-manager-tests
Change-Id: Ied4755a22cace2047b00d388292fafb610c921e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630209
Reviewed-by: Onath Dillinger <claridge@google.com>
Reviewed-by: Didi She <didis@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>