[graphics][compute][spinel2] Add BGRA8 opcode to ease testing
Using a compile-time swizzle to select the BGRA8 ordering on host
tests was cumbersome and complicated the the unit testing framework
(`spinel2-vk-test`) and benchmark (`spinel2-vk-bench`).

Instead, an host application can select to store with either an RGBA8
(Fuchsia) or BGRA8 (host) format before optionally copying to an
image.

Long-term, the BGRA8 opcode will be removed along with the internal
copy-to-swapchain and replaced with a tile-copying vertex>frag
sequence.

Testability:

 * Ran unit tests against both K2200 (sm35) and RTX 2060 (sm75):

   out/core.chromebook-x64/host_x64/exe.unstripped/spinel2-vk-test

Change-Id: If60f90e46a7255b205a4fbf3b55c40d2230fd8a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638602
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>