[fidl][cpp] Initial natural API test
This is the start of tests for the natural API bindings. So far it just
contains a copy of the wire API tests for enums and bits with the symbol
names changed - since those are the same types but in a different
namespace.

This also moves the fidl.llcpp.types.test FIDL library definition out of
the wire_types subdirectory since it's used for more.

Change-Id: I8563c835ab8abe6c1bf5bc4066661de82a97fe6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635537
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>