[input][tests] Create a (WIP) workstation test realm
This is a static definition of a workstation test realm. It can be used
from within a RealmBuilder to create a sub-realm that replicates the
static workstation setup for tests.

Note that, sadly, the RealmBuilder seemingly does not allow loading CFv1
components from within such a manifest file, so a number of components
had to be wrapped into a dedicated runner. Since this is expected to be a
one-time setup, it should not complicate the reasoning about the realm
very much.

Bug: 92183
Change-Id: I773a288bbee0a4fb60e4da3e9c2216e480cfe258
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638675
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>