[run_test_suite] Stop polling run events on cancel
Since it's possible for test_manager to hang for some reason
serving run events, we should cancel retrieving these events as
well as the suite events.
This is intended to address flakes, but in the future we might want
to poll run events again so that users still get artifacts if a test
is stopped prematurely.

Bug: 92308
Multiply: run_test_suite_integration_tests

Change-Id: Ic0393c9bcaab69c479753768ffe16fff73351f60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638667
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>