[zircon][hpet] Change the way we compute the nominal HPET frequency.
Change the way that we compute the nominal HPET frequency (and
therefore, the HPET ticks to clock monotonic ticks ratio).

HW reports the nominal HPET timer _period_ in a register expressed in
femtoseconds.  According to Intel docs, this value must imply a
nominal frequency of at least 10MHz, but could (technically) imply a frequency
as high 1 petahertz.

The problem with this approach, is that the actual nominal HPET rate
is typically equal to the frequency of the main motherboard crystal
(24MHz, 100MHz, etc).  When the rate is 100MHz, the ratio between the
HPET and the (virtual) clock monotonic rate (1GHz) is a simple 10:1.
If the rate is 24MHz instead, the ratio _should_ end up being 125:3.
Because of the way that the ratio was being computed from the period,
however, we were ending up with a rate of 41666667/1000000.

Typically, this is not really a problem, at least not until we start
to combine this ratio with other ratios (such as the conversion used
to perform clock correction for UTC).  Clock correction ratios are
usually expressed in PPM values (for example, 1000017/1000000).  From
a ratio perspective, 41666667 is a particularly nasty number.  Its
prime factors are 3,7,109,109,167.  1000017 (the value used in this
example) has the factors 3,3,23,4831.  Note the lack of any 2s or 5s
in these numbers.  The product of these ratios ends up being something
which can no longer be stored in 32 bit ratio, which was causing an
ASSERT in libaffine to fire during clock updates on an Intel NUC
because of the inability to store the composite ratio precisely.

So, there are really two issues here.  First is the ASSERT in the
clock update.  The proper thing to do here is to allow libaffine to
use an inexact reduction of this ratio, thereby preventing the assert.
Clock conversions will be slightly less precise, but this should be OK
in the long run (since the clock is being actively maintained).  This
will be addressed in a separate CL.

Second is the fact that the numbers being used for nominal conversions
were technically incorrect.  The nominal rate of the HW I have is
actually 24MHz, not 23.999 as was being reported (it is easy to tell
since the APIC on this device is _also_ running at 24MHz, and the TSC
is running at the CPU issue rate of 2.4 GHz).  So, instead of allowing
this imprecision to creep in, we approach the calculation of the
nominal frequency slightly differently.  When computing the nominal
frequency from the period, we assume that the rate is an integer
number of hertz (as it normally is) by rounding up to the nearest
number of Hz during the division instead of rounding down.  Nothing
changes when our HPET rate is 100MHz, but we end up with the correct
number in the case that the HPET is running at 24MHz.

Tests: existing unit tests, and manual testing on KVM QEMU as well as
       NUC.
Change-Id: I27d9fa5507723851f4162c446471d38a314f09ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637678
Reviewed-by: Jody Sankey <jsankey@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: John Grossman <johngro@google.com>