[flatland] Set Clips in Local Transform Space
Make it so that transform clip regions are specified
in the local coordinate space of the transform node
that they are attached to. Internally to Flatland,
these local clip regions are all converted to a global
space to perform clipping calculations.

Updated tests in global_transform_data_unittest.cc

Change-Id: I52d03af958396035fe14cb2b200e12c24e4aca0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636841
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Chris Fontas <cfontas@google.com>