[ffx] Fix the --debugger flag
Code review and testing brought up a problem with how we're handling the
debugger flag in ffx emu. When executing `ffx emy start --debugger`, we
expect the emulator process to pause before booting Fuchsia so that the
user can attach a debugging process like gdb first. Instead, the
launcher process would pause, but the emulator would just start up as
normal. If the user added the --console or --monitor flag, debugging
would be ignored.

This CL adds the appropriate -wait-for-debugger flag to the emulator
flag template, predicated on the debugger flag of course, so the
emulator will actually pause like it's supposed to. We also remove the
code that was causing the launcher to wait for the emulator to return,
as this was a carryover from device_launcher and doesn't fit our
existing execution model anymore.

Now, the launcher process will exit once the emulator is running (or
branch to the console/monitor, if needed), and the emulator process will
pause, as expected.

Change-Id: I9d1d0a46567052e16bf551f5ff5637526d9000cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638724
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>