[fidl][cpp] Natural event sending
Adds |fidl::SendEvent| for sending events with natural types.

It uses similar structure as |fidl::WireSendEvent|, but is much simpler
because there are no caller-allocated flavors.

The return type is |fitx::result<fidl::Error>|, which is the same as
that of one-way client calls.

Now we have everything that's needed for a pair of end-to-end async
clients and servers.

Added a test in the wire_interop_tests.

Bug: 60240

Change-Id: I3b79253433d72075142b7d68fc97dc36b9daaf79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638203
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>