[fdf][mock-ddk] Allow device_add to take nullptr
Currently, device_add_from_driver allows the |out|
argument to be null, in the situation where a driver
does not need to maintain a pointer to the device added.

This adds that behavior to the mock-ddk and the fake-ddk.

Test: Included.
Change-Id: I713393be68ac696201b1b4ebfb68519eabe547c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638679
Fuchsia-Auto-Submit: Garratt Gallagher <garratt@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>