[routing] Use PartialChildMoniker for comparisons
`find_matching_offer` and `target_matches_moniker` functions take a
`PartialChildMoniker` instead of a `ChildMoniker`, because their
comparisons do not use or need the `InstanceId` field information from
`ChildMoniker`.

The reason for this change is to reduce unnecessary uses of monikers
with instance ids.

Bug: 87407
Change-Id: I12de4988bd06e3ad794c5658ff9106b00ac76608
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638808
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>