[ermine] Add logout button
This change adds a button to logout of the ermine shell. When oobe is
enabled, the user is sent to the login screen. When it is disabled, the
user is logged out and then logged back in with an empty password.

The buttons are updated to icon buttons to confirm to latest ux design.
- Adds a helper dart extension to add tooltip to any widget.

Bug: 92301
Change-Id: I20c1b32f2542dfceb385082c828ff8adfcc78da7
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/638981
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>