[system-update-checker] Check for empty target channels
In https://fxrev.dev/627872, we switched system-update-checker to use an
empty string if we didn't find the target channel in vbmeta.
Unfortunately this caused us to try to ota with `fuchsia-pkg:///update`.
This patch checks for this condition when trying to construct the update
url, and returns `None` if the target channel is empty. This causes us
to fall back to the `fuchsia-pkg://fuchsia.com/update` url.

Fixed: 91886
Change-Id: Ia6977182ff99f3a2ea5ef456655f2a9492b255b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635373
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>