[cts] FIDL Signature tests
This CL is the first step towards running signature tests in the CTS.
This creates the python script to parse the manifest file and run the
diffing programs, and creates a build template that pulls everything
together.

The PlaSA manifest file and diffing tools will be added to the CTS
archive after this CL is submitted. Future CLs will copy over the
fragment files, and actually execute the plasa_differ_release test as
part of each CTS release.

Change-Id: I8f3642174ee4cdd458c65bedc5075b1d1a915426
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624326
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: John Shamoon <johnshamoon@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>