[assembly] Use PackageBuilder to make packages in assembly
Switches the creation of the 'base' and 'update' packages over to using
the PackageBuilder.

Adds the ability to add files created from passed-in content to the
builder.

Adds the ability to set the ABI revision to a package.

Removes the consolidated create_meta_package_file() fn from the
utils lib, as it's now part of the PackageBuilder.

Bug: 91743

Change-Id: I3a137e93c8518f0fa9407b26700973198eed064a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637664
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>