[ermine] When dismissing views, move focus to previous view instead of next
This is important for Chromium and Linux apps where it is
confusing to the user when focused is moved to the next
view instead of what had focus previously.

Test: Launch two terminals and one spinning square, close
spinning square, focus is returned to the last focused
terminal instead of the first.
Bug: 91433

Change-Id: I5e523f5829ecd85d54b59012899b9ae09455998f
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/639601
Fuchsia-Auto-Submit: David Reveman <reveman@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>