[Workstation] Switch to component manager bootfs
This flag stops creating bootfs in bootsvc/C++, and instead creates it
in component manager/Rust. While this is a well tested change,
workstation is a good first product to enable this for as workstation
is still unreleased.

Bug: 44784

Tested:
Validated we're using the right bootfs from CQ logs

Change-Id: I515b78e4798b6a5519f9fe5ccfa0d7adf46bbd04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638401
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: David Hastings <dahastin@google.com>