[netemul] Define and parse network config
Until the component manifest schema supports it, the network
configuration for a netemul runner test will be specified in a separate
JSON file bundled with the test package.

Also, this CL does not introduce the actual configuration logic that
will exercise the netemul sandbox and the configurable netstack; it just
introduces the configuration schema and reads and validates it in the
runner.

Fixed: 92175
Test: fx test netemul-runner-unit-tests
Test: fx test netemul-runner-tests
Change-Id: I01d610e05a88ff85f7047d5855a1791b7b91b8c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638601
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>