[accountctl] Print names for each account on the device.
If the use different names in different account creation flows this will
be a useful debugging aid. If not then it will still be useful at some
point in the future.

Note that account-ids doesn't make much sense as a subcommand if it
does both IDs and names so I changed to "list". I'm enjoying the fact
that thats easier to type!

Change-Id: If05af69783b04d92904f6c4d3d2b9285d894a1ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638726
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>