[utest] Reland "Move object property syscall tests into core-tests"
This is a reland of fee712ea318f98d89f8cd58d9060d415821e7137.
The test instability is introduced by the old property test's
busy-wait thread consuming a whole CPU for the rest of the
standalone test.  That is resolved by the revamp of that test
code in a follow-on change that will land together with this
reland.

Original change's description:
> [utest] Move object property syscall tests into core-tests
>
> Change-Id: I4d239f20cc979483d011fb12ccd9ed44e5322f4c
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638639
> Reviewed-by: Nick Maniscalco <maniscalco@google.com>
> Commit-Queue: Roland McGrath <mcgrathr@google.com>
> Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>

Change-Id: I20cbe44e694dcea8ad03223de1168920f13b705d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639575
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>