[kernel] Allow ZX_PROP_REGISTER_{FS,GS} in zx_object_get_property, improve test
This adds a much more thorough test for the behavior related to
the %fs.base and %gs.base registers.  Partly to facilitate this
test, and partly just to make ZX_PROP_REGISTER_{FS,GS} a more
complete software substitute for the newer CPU instructions,
zx_object_get_property now accepts these properties just as
zx_object_set_property does, so it can provide a software
replacement for rdfsbase/rdgsbase as well as wrfsbase/wrgsbase.

The new test exercises the CPU instructions as well as the system
calls, so it can verify that they interact consistently.  To test
both with and without the CPU instructions available, QEMU (with
or without KVM) can be passed the `-cpu -fsgsbase` switch when
running on hardware that does actually support it to emulate
hardware that does not.

Change-Id: Ice4e436e66dee532cc75ec185c58e103fb4a14e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639371
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>