[core_size_limits] Set absolute limits
The size limits were set to an arbitrarily high limit before. This
change modifies the size limits to be high, but within the realm of
reason so that the numbers won't look ridiculous once we show them in
the binary size table.

The size limits were determined by taking max(core.arm64, core.x64) for
each component's current byte consumption, multiplied by 1.5, so there
is plenty of room to grow. The intent remains that we do not wish to
place a cap on the core product's components, only that we want to use
the creep limits to make organic growth more intentional.

The next step to this will be to add core_size_limits.arm64 and
core_size_limits.x64 to the binary size table in Gerrit, but as FYI
only.

Bug: 83734
Change-Id: Ida7379e408a1d5e3f3a6ad25a73ae6f3403f749d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639580
Reviewed-by: Saman Sami <samans@google.com>
Commit-Queue: Anthony Fandrianto <atyfto@google.com>