[build] Allow actions to specify ignore prefixes for action_tracer
This allows actions to specify a list of prefixes which the action tracer
will ignore.  This is useful for mimicking directory-glob semantics where
an action needs to recursively delete an entire directory tree in the
outdir, but on an incremental rebuild, the contents of the directory tree
may be different.

Fixed: 92503

Change-Id: I57ccef76b1706240693c2eca33c77eff9bd6e6c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639577
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>