[moniker] Rename AbsoluteMoniker to InstancedAbsoluteMoniker
`AbsoluteMoniker` (the moniker representation inclusive of instance ids)
is renamed to `InstancedAbsoluteMoniker`. Struct fields and variables
named `abs_moniker` that point to an `InstancedAbsoluteMoniker` are
renamed to `instanced_moniker`.

This CL is a rename only.

The reason for this CL is to isolate uses of formerly named
`AbsoluteMoniker` away from the public API space.
`InstancedAbsoluteMoniker` only becomes visible to component_manager.

Bug: 87407
Change-Id: I9baedcc6e6293c9b290f36681aa9408c3f11a1b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639044
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>