[workstation_pro][chrome_smoke] Added screenshot test
Improving workstation_chrome_smoke_test by making it take a screenshot to see if a blank white page of the Chromium view is rendered.
The original test failed to catch a crash because Ermine seemingly succeeded to present a Chromium view but just failed to render it due to the crash. Screencapturing is to prevent false positive cases like this one.
Ideally, we should enable workstation_chrome_advanced_smoke_test, which runs a local test server and checks if a certain page served by the server is loaded, but it needs to be deflaked before being moved to CI.
In the meantime, enhancing this basic smoke test so that it can avoid false passing as much as it can.

Tested on a local machine 10 times and all passed.

Bug: 92206

Change-Id: I90881bb8e1c0d472f63240ba87575ccd46339007
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/639343
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Yeonhee Lee <yhlee@google.com>