[moniker] Rename PartialAbsoluteMoniker to AbsoluteMoniker
`PartialAbsoluteMoniker` (the moniker representation that does
not include instance ids) is renamed to `AbsoluteMoniker`. Struct fields
and variables named `partial_abs_moniker` are named to `abs_moniker`.

This CL is a rename only.

The reason for this CL is to normalize the use of formerly named
`PartialAbsoluteMoniker` as the only moniker representation available
for future development.

Bug: 87407
Change-Id: Iacddf862d85fbdbb135c82121dc5a139a4630511
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639045
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>