[assembly] call product assembly during the build
This allows the product assembly code to be exercised, and allows
the migration of features from assemble_system() to product
assembly, for use out of tree.

The image assembly configuration generated by product assembly
is validated against that created "the old way".  This is a
check to ensure that an equivalent configuration is being created.

Bug: 91743
Change-Id: I56d97fcff60be6b17c406e60e4c632b4f71898ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634546
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>