[inspect][codelab] Make part 1 and 2 of codelab more consistent
The codelab had some inconsistencies that were introduced with the
migration to CFv2. This change removes the inconsistencies in part 1 and
2 and fixes some typos.

Additionally since CFv2 the FizzBuzz component for part_2 of the codelab
needs to be added via the component Binder otherwise it's not started.

Change-Id: I0ae5b6d0e8886f9409a3e2d4b032e663845fd843
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636501
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Adrian Hausner <adrianhausner@google.com>