[lsc] Migrate all Realm Builder clients to new namespace
This LSC migrates all clients of the Realm Builder C++ library
to use the correct namespace, `component_testing`. Prior to this
change, they were using a combination of types/functions exported
in the `sys::testing` and `sys::testing::experimental` namespaces.
This concludes the in-tree of migration of the C++ library Realm
Builder library to the new version outlined in RFC 140.

Bug: 90794
Change-Id: I32775f1b0765895fa57dabe43ad7b4e29c1030bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639673
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>