[ZBI Parser] Slightly tweak the get item API.
The item services use the extra field as a key, so we need to include
the extra field in the ZBI parser result.

This is also generally more extensible, as we could now trivially add
more metadata to the result as needed.

Bug: 44784

Tested:
Unit test

Change-Id: Ica45fa6ad28bd4f4dd0d88b5d07be18288bebc3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639926
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Hastings <dahastin@google.com>