[ermine] Move Ermine to the account storage capability.
The long term plan was always that components inside a session should
persist their data on the "account" storage capability. These
components potentially deal with sensitive user data so it is
important to encrypt these data at rest using the user's password,
which is what the "account" storage capability does.

Bug 92479 in CF actually means the "data" storage capability isn't
currently working inside a session, so moving over is more urgent.

Note this also adds a single startup log lines for ermine to report if
it found settings or not. Happy to remove this if needed.

Bug: 92479
Change-Id: Ibe724ea2b7ca03a5e43b6d156d3229cc9dcb6daa
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/639765
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>