[reachability] Rework "Gateway" criteria
Consult the route table to collect routes that have a "next hop" and
treat those addresses as gateway addresses. This is in contrast to the
previous logic that incorrectly considered assigned addresses' subnet
mask in attempting to determine if a suitable local address can be
selected for the route; this is the responsibility of the network stack,
so can be delegated to it.

Ping gateways concurrently while I'm here.

Change-Id: Id695eb0b1a8c77a06f47966221d9cd1d729c6b7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640142
Reviewed-by: Devon O'Dell <dhobsd@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>