[mold] Abstract buffers.
This patch updates the way buffers are used in mold. It aims to abstract over
the possible layouts that the Buffer type can have. For the moment, only
LinearLayout is implemented but it will allow us to implement other layout
types in the future.

Every layout type has a TileWriter associated with it that will be in
charge of writing in the buffer.

When rendering, we iterate over each tile and create an associated
TileWriter instance depending on the layout.

This patch introduces:
- a new type, SplitsCache, that encapsulates cached buffer pointer usage
  in a safe wrapper
- a Layout trait to create new layout types (e.g LinearLayout)
- a TileWriter trait used by the Layout to write to the buffer (e.g
LinearTileWriter)

Change-Id: Id50356f786eda54a756bfda1468db27ac424fec8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613647
Reviewed-by: Dragoș Tiselice <dtiselice@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: David Reveman <reveman@google.com>