[netstack3] Mock timer methods take callbacks
Previously, the `DummyTimerCtxExt::trigger_xxx` family of methods
required that `Self` implemented `TimerHandler`, and operated by calling
`TimerHandler::handle_timer` for each triggered timer. In this change,
those methods are modified to accept a callback which is used instead of
`TimerHandler::handle_timer`.

This makes it so that these methods can be used by code which does not
use `TimerHandler`.

Bug: 92605
Change-Id: I6d9e0966ed88bdcde2e0bc2ffd7d008abb92258d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640421
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>