[cm_moniker] Rename ExtendedMoniker to InstancedExtendedMoniker
This CL renames ExtendedMoniker to InstancedExtendedMoniker to signify
that it contains a ComponentInstance variant with an
InstancedAbsoluteMoniker. This continues to live in the cm_moniker
library; it is used by component_manager and allowed dependencies that
require an instanced absolute moniker from an extended moniker.

The type ExtendedMoniker is available in the moniker library; this type
includes a ComponentInstance variant that contains an AbsoluteMoniker
without instance ids. This is the preferred type to use for development.

This CL is a rename only.

The reason for this CL is to isolate the areas that require an
InstancedExtendedMoniker (primarily consumers of the event system), from
areas that do not need an instanced moniker variant from extended
monikers (to be addressed in a future CL).

Bug: 87407
Change-Id: I5a51b5068b0ded30016c0d10c26c405d4ba3d603
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639988
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>