[ffx] Update argh documentation for ffx emu
As we move toward releasing the `ffx emu` tool to general use, we need
to update documentation to make it easier for the new users to know what
to do. This CL starts this process by adding details to the help-text
for subcommands and their args, which is also propagated to fuchsia.dev
and will serve as the starting point for other doc changes moving
forward.

Change-Id: I54e68274099c45c119e5008127cfc71ff26c81fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640122
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>