[ermine] Add logout confirmation alert dialog
This change adds a confirmation dialog when the user invokes logout,
restart and shutdown.

The dialogs code has been refactored to use Flutter's Navigator widget
to manage multiple dialogs show on the screen and automatically dismiss
them when the user presses Escape key.

This change also displays the dialogs on top of child views, since the
user can invoke logout when a child view is displayed fullscreen.

After the refactor, only AlertDialogs can be displayed. This can be
extended in the future.

Bug: 92301
Change-Id: Ie4cfe5e70c621039688b071cbf969baf00f02fa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/640244
Reviewed-by: Charles Whitten <cwhitten@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>