[ffx] Use the --log flag to redirect QEMU output
In ffx emu, currently we just inherit stdout/stderr for the femu
and qemu binaries. This gets really sloppy on the terminal, especially
after ffx terminates and femu/qemu keeps logging over the top of the
terminal prompts.

This CL addresses that problem by redirecting femu/qemu output to a
user-defined log file instead. That file is specified through the --log
flag, and if no log file is specified, it goes to a default location in
the emulator's state directory.

This makes for a much cleaner terminal, while also allowing user's to
more easily revisit or share emulator output when needed.

Change-Id: I8b2b2dd87fd9c042591a4ed2028ae830884f48b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634221
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>