[netstack3][pmtu] Split contexts
This is part of a larger change in which all context arguments in the
Core are being split in two in order to resolve some lifetime conflict
issues and to prepare for a deeper overhaul of the way state is accessed
and synchronized. See https://fxbug.dev/92605 for more details.

Bug: 92605

Change-Id: I88752e9f18ae43540cf7932b0fdca2502ff9312f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639789
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>