[component_manager][component_tree_stats] Stats keys on non-instanced monikers
This CL has diagnostics stats use monikers without instance ids to key
on for component instances.  `InstancedAbsoluteMoniker` and
`InstancedExtendedMoniker` are replaced with `AbsoluteMoniker` and
`ExtendedMoniker` respectively.

When the Started event triggers the `on_component_started` hook, the
hook will not return early if the stats tree already contains the
moniker, and a `diagnostics_waiter_task` will get spawned for the
(duplicate) moniker. This results in a task measurement created for two
dynamic instances that share the same `AbsoluteMoniker` to share the
same key in the component stats tree.

This change depends on a Started event having a 1:1 mapping with a start
action (ie there cannot be multiple started events dispatched from one
start action).

The reason for this change is to reduce the use of monikers that include
instance ids in business logic where possible.

Bug: 73171
Change-Id: Ib543a2be92b678ccac53b88b5061fe1c9b2ed3e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640154
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>