[fxfs] Fix journal corruption
If journal replay advances the file offset to a non-block-aligned
offset, the writer was erroneously configured to start writing entries
immediately at that offset, rather than the next block-aligned offset.

This is in incorrect, because then the journal writer has to
read-modify-write existing journal blocks, which can result in
corruption since the block writes might not be atomic and existing
serialized records might end up in an indeterminate state.

Fixed this so that the journal writer always skips to the next
block-aligned offset.  This also has the benefit of preventing the
journal writer from doing costly read-modify-writes.

Re-enables `fxfs-crash`, which reproduces this issue.

Bug: 91684
Fixed: 92435
Change-Id: Ifd7e721b7efa9dfa9b25cba6c062b2b38962748c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641041
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>