[netstack3][reassembly] Split contexts
This is part of a larger change in which all context arguments in the
Core are being split in two in order to resolve some lifetime conflict
issues and to prepare for a deeper overhaul of the way state is accessed
and synchronized. See https://fxbug.dev/92605 for more details.

Bug: 92605
Change-Id: Ie6726d1f1036ff43c40ecbf683351224d9d1caf6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639790
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>