[archivist] Distinction between external and internal event producers
This change introduces two types of event producers:
1. Internal: emit events that are purely internal, they may or may not
   originate from external actions.
2. External: emit events that come from external sources (such as v2
   event source, v1 component event provider, unattributed log sink
   connections, etc).

The goal of this is to provide a mechanism for shutdown. During shutdown
we perform the following actions before ensuring all logs are drained:
1. Stop external event ingestion
2. Drain all external events that were already buffered.

This change provides the the API to accomplish that, the following
change will exercise it in the archivist.

Bug: 92016
Multiply: archivist-unittests, archivist-integration-tests-v2

Change-Id: I23bf95df93e2605c543ac55e2d58dc08224443bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640123
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>