[archivist] Use new event router
This is a first step towards the event bus architecture for the
archivist. In this CL we refactor the existing event ingestion to use
the new API.

A lot of the advantges of the new API aren't still clear since we still
don't decouple repositories and have them emit events too. The event
router producer and consumer installation happens in diverse places in
archivist.rs, the goal is to bring them to the same spot so that it's
very clear how the connections are made.

Multiply: archivist-unittests, archivist-integrationt-tests-v2
Fixed: 92016

Change-Id: I49b139510a8ee126dde2e27b0a7a669f8fa133be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635924
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>